[tmux] Load tmux statusline config from neovim
Previously, tmux would read ~/.config/tmux/tmuxline.conf for the
statusline configuration. This was a symlink to one of
~/.config/tmux/tmuxline.conf.{light,dark}, which were in turn files
written by the tmuxline.vim plugin.

This is a brittle setup, since if we change the neovim statusline
configuration, then we manually have to re-generate the
tmuxline.conf.{light.dark} files.

However, the tmuxline.vim plugin will also set the statusline for tmux
whenever we open neovim. We can make use of this, by launching and
quitting a neovim instance when launching tmux. This way, the tmuxline
configuration is now set by tmuxline.vim directly, instead of indirectly
via tmuxline.conf.{light,dark}.

With this new solution, the tmux statusline always stays in sync with
the neovim statusline.